Add /killall, misc fixes

+++ src/main/java/me/pugabyte/nexus/utils/EntityUtils.java
@@ -12,12 +12,12 @@ import java.util.stream.Collectors;
 
 public class EntityUtils {
 
-	public static LinkedHashMap<Entity, Long> getNearbyEntities(Location location, int radius) {
+	public static LinkedHashMap<Entity, Long> getNearbyEntities(Location location, double radius) {
 		return Utils.sortByValue(location.getWorld().getNearbyEntities(location, radius, radius, radius).stream()
 				.collect(Collectors.groupingBy(Function.identity(), Collectors.counting())));
 	}
 
-	public static LinkedHashMap<EntityType, Long> getNearbyEntityTypes(Location location, int radius) {
+	public static LinkedHashMap<EntityType, Long> getNearbyEntityTypes(Location location, double radius) {
 		return Utils.sortByValue(location.getWorld().getNearbyEntities(location, radius, radius, radius).stream()
 				.collect(Collectors.groupingBy(Entity::getType, Collectors.counting())));
 	}
